move GuidedSetup to its own component in scenes
