+++ src/scenes/GuidedSetup/styles.js
